requirements.txt
pytest
requests
allure-pytest
allure-python-commons
faker